Revert "refactor: header & footer rewrite"
This reverts commit f626cf20c3cbfb3869d2aa2e4a2113523cd7577e.
