+++ src/components/Footer/Footer.css
@@ -3,7 +3,8 @@
     left: 0;
     bottom: 0;
     width: 100%;
-    height: 60px;
+    height: 100%;
+    max-height: 60px;
     text-align: center;
 
     display: flex;
