+++ src/components/Footer/Footer.js
@@ -1,5 +1,5 @@
 import React from "react";
-import './index.css'
+import './Footer.css'
 
 function Footer() {
     return(
