--- src/components/header/style.css
@@ -1,20 +1,0 @@
-.header-navbar {
-    height: 80px;
-    width: 100%;
-    top: 0;
-
-    background-color: #DFF2ED;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-    margin: 0px;
-    padding: 20px; 
-    list-style-type: none;
-    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-    border-radius: 0px 0px 16px 16px;
-}
-
-.header-login{
-    margin: 30px;
-}
