+++ src/pages/Customer_Care/CustomerCare.js
@@ -2,7 +2,7 @@ import React from "react"
 import { Routes, Route } from 'react-router-dom';
 import { Container, Row } from "react-bootstrap";
 import './CustomerCare.css';
-import Footer from '../../components/Footer';
+import Footer from './components/Footer/Footer';
 import Header from './components/Header/Header';
 import CSKHBanner from "./components/CSKHBanner/CSKHBanner";
 import ContactUs from "./components/ContactUs/ContactUs";
