+++ src/pages/TicketBooking/components/Filter/index.css
@@ -1,5 +1,5 @@
 .filter-box {
-    padding-top: 30px;
+    padding-top:240px;
     position: sticky;
     top: 0;
 }
